requirements.txt
django
crewai
python-dotenv
crewai[google-genai]